pyproject.toml
[tool.poetry]
name = "calendar-service"
version = "0.1.0"
description = ""
authors = ["arsar7 <sargsyan_arman@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.7.4"
requests = "^2.32.3"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
greenlet = "^3.0.3"
aiosqlite = "^0.20.0"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
